+++ addons/common/stringtable.xml
@@ -37,7 +37,7 @@
             <English>Doctor</English>
             <German>Arzt</German>
         </Key>
-        <Key ID="STR_ttt_Common_EOD">
+        <Key ID="STR_ttt_Common_EODunit">
             <English>EOD Technician</English>
             <German>EOD Spezialist</German>
         </Key>
@@ -48,6 +48,7 @@
         <Key ID="STR_ttt_Common_EngTL">
             <English>Engineer Team Leader</English>
             <German>Pionier Truppführer</German>
+        </Key>
         <Key ID="STR_ttt_Common_vba">
             <English>TTT Medical Crate (VBA)</English>
             <German>TTT Sanitätskiste (VBA)</German>
